+++ actvfunc.h
@@ -5,11 +5,11 @@
 #ifndef NEURALNETWORK_ACTVFUNC_H
 #define NEURALNETWORK_ACTVFUNC_H
 
-float Sigmoid(float x);
-float d_Sigmoid(float x);
-float ReLU(float x);
-float d_ReLU(float x);
-float Leaky_ReLU(float x);
-float d_Leaky_ReLU(float x);
+double Sigmoid(double x);
+double d_Sigmoid(double x);
+double ReLU(double x);
+double d_ReLU(double x);
+double Leaky_ReLU(double x);
+double d_Leaky_ReLU(double x);
 
 #endif //NEURALNETWORK_ACTVFUNC_H
